# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MetaID Indexer API",
    "description": "Multi-chain blockchain indexer API for MetaID protocol. Indexes PIN inscriptions and MRC20 token operations across Bitcoin, Dogecoin, and MicroVision Chain.",
    "version": "2.0.0",
    "contact": {
      "name": "MetaID Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/pin/{numberOrId}": {
      "get": {
        "summary": "Get PIN by ID",
        "description": "Retrieve a specific PIN by number or ID",
        "parameters": [
          {
            "name": "numberOrId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      }
    },
...
  },
...
}